ci: Move checksum stuff into its own module
Just a little cleanup.
